Tool picker is not switch statement
Previously, tool picker was in many if statements based on many bools. Now, one 'toolID' Decides which tools is active. 0 is no tool, and other integers represent other tools. It would be easy to add new tools, as they would receive new integer values to represent them.
This is good because keeping track of the many bools is harder than having one int value which can either be valid or invalid.