Added a File Clear to the Load Function
The issue reported about file upload. The file will always be uploaded there is no way to confirm the file to be an image without uploading it. All we can do is clear the value and not load the image. The function is already not loading without a match, so I added the clear, hopefully.