Added cap to image size and download button
Not sure if this is a permanent change, but now if a image uploaded is too large to fit on our screen, it is invalid. The image can only be some fraction of the size of the screen, as determined by maxX and maxY.
Additionally, a button and function have been added that download the finished image. Currently, this opens a new page. This is not a bug, but could be changed later to look better.