Release 2.3
Prevented tool radios change while using a tool, which previously caused a bug.  Added preview brush effects for all brushes. Reduced opacity of brushes which had to have very low opacity to be noticeably different. Double width of pencil tool, where line width was size, but the size was supposed to be diameter.